I ended making the static ArrayList<Ammo> in Ship
I don't know if it fits Java convention but considering each ship will
need to check this collision anyway it might work out... in the future
Terrain needs to collide with bullets but that can be handled by the
Game or Map class.

I don't know let's see if it makes sense, I'll fix it later if it
doesn't. Lol.